# openapi.yaml
openapi: 3.1.0
info:
  title: Market Data API
  version: 1.8.0
  description: Consulta cotações diárias não ajustadas publicadas pelo B3 Data Hub.
servers:
  - url: /
paths:
  /health/live:
    get:
      operationId: getLiveness
      summary: Verifica se o processo HTTP está vivo
      description: Não consulta PostgreSQL nem outras dependências externas.
      responses:
        '200':
          description: Processo saudável
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HealthStatus' }
  /health/ready:
    get:
      operationId: getReadiness
      summary: Verifica se a instância está pronta
      description: Executa um Ping leve no PostgreSQL com timeout configurável. Não consulta dados de negócio.
      responses:
        '200':
          description: Dependências obrigatórias disponíveis
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HealthStatus' }
        '503':
          description: PostgreSQL indisponível; nenhum detalhe interno é exposto
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HealthStatus' }
  /metrics:
    get:
      operationId: getOperationalMetrics
      summary: Retorna métricas operacionais acumuladas pela instância
      description: Métricas JSON em memória, reiniciadas com o processo. Rotas e operações usam dimensões de cardinalidade controlada.
      responses:
        '200':
          description: Snapshot das métricas HTTP, PostgreSQL e runtime Go
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OperationalMetrics' }
  /assets/{ticker}/intelligence:
    get:
      operationId: getAssetIntelligence
      summary: Retorna indicadores determinísticos de um ativo
      description: >-
        Consulta somente leitura com timeout de cinco segundos. Rejeita parâmetros
        desconhecidos, repetidos ou vazios. asOf limita a data de pregão usando a
        versão atualmente publicada dos dados, não a versão conhecida no passado.
        Sem asOf usa a última cotação até hoje em São Paulo. A referência efetiva
        pode anteceder a solicitada. Preços não ajustados. Janelas d são pregões.
        Usa calendário observado consolidado do COTAHIST, com integridade de cada
        importação validada e origem/versão em meta.calendar quando includeDetails=true. Não equivale a um
        calendário oficial independente: omissões globais na fonte não são detectadas.
        Anos ausentes ou cobertura inválida indisponibilizam somente as janelas afetadas.
        Sessões são limitadas à última cotação até asOf. O percentil histórico do RSI
        usa anos civis, exclui o RSI da própria referência e exige 252 observações válidas.
        Quando benchmark é informado, ativo e benchmark são carregados no mesmo snapshot
        e comparados nas últimas oito sessões consecutivas do ativo quando todas também existem para o benchmark.
        Os sinais são avaliações técnicas determinísticas dos indicadores e não constituem recomendação de investimento.
        Retorna somente indicadores implementados. Status complete significa que todos foram calculados; caso contrário, partial.
      parameters:
        - $ref: '#/components/parameters/Ticker'
        - $ref: '#/components/parameters/MarketType'
        - name: includeDetails
          in: query
          description: Inclui meta.calendar, meta.calculation_version, meta.ruleset_version, meta.data_version e meta.rsi_seed_from.
          schema: { type: boolean, default: false }
        - name: rsiWindow
          in: query
          description: Janela civil do percentil histórico do RSI(14), expressa como um número positivo de anos seguido por y.
          schema: { type: string, pattern: '^[1-9][0-9]*y$', default: 3y, example: 5y }
        - name: benchmark
          in: query
          description: Ticker opcional usado para retorno comparável e força relativa. Deve ser válido, ter cobertura na fonte e ser diferente do ativo.
          schema: { type: string, pattern: '^[A-Za-z0-9]{4,12}$', example: IBOV }
        - name: asOf
          in: query
          description: Data civil não futura e diferente de 0001-01-01.
          schema: { type: string, format: date, example: '2026-09-04' }
      responses:
        '200':
          description: Análise completa ou parcial; null representa indisponibilidade, nunca zero.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IntelligenceEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
        '504': { $ref: '#/components/responses/RequestTimeout' }
  /comparisons:
    get:
      operationId: compareQuotesGet
      summary: Compara o desempenho de dois a dez ativos
      description: >-
        Consulta somente leitura, com período inclusivo de até cinco anos e timeout de cinco segundos.
        Tickers são normalizados e duplicidades são rejeitadas. Parâmetros desconhecidos ou
        repetidos são rejeitados. Preços não são ajustados por eventos corporativos.
        Pelo menos dois ativos devem ter dois fechamentos positivos em datas distintas.
        Volatilidade amostral anualizada exige três cotações; correlação exige dois intervalos
        de retorno coincidentes e variância positiva. Métricas indisponíveis são omitidas.
      parameters:
        - name: tickers
          in: query
          required: true
          style: form
          explode: false
          schema:
            type: array
            minItems: 2
            maxItems: 10
            uniqueItems: true
            items: { type: string, example: PETR4 }
          example: [PETR4, VALE3]
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
        - $ref: '#/components/parameters/MarketType'
        - name: metrics
          in: query
          style: form
          explode: false
          schema: { $ref: '#/components/schemas/ComparisonMetrics' }
        - name: benchmark
          in: query
          description: Ticker da mesma fonte; resumo separado, sem participar automaticamente dos rankings. Sem dados, retorna insufficient_data.
          schema: { type: string, example: BOVA11 }
        - name: includeSeries
          in: query
          schema: { type: boolean, default: false }
      responses:
        '200': { $ref: '#/components/responses/ComparisonSuccess' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '422': { $ref: '#/components/responses/InsufficientComparisonData' }
        '500': { $ref: '#/components/responses/InternalError' }
        '504': { $ref: '#/components/responses/RequestTimeout' }
    post:
      operationId: compareQuotesPost
      summary: Compara ativos a partir de um corpo JSON
      description: >-
        Consulta somente leitura com as mesmas regras e resposta do GET /comparisons.
        Não persiste a comparação. Aceita um único objeto JSON, sem campos desconhecidos,
        com tamanho máximo de 64 KiB. Timeout de cinco segundos.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ComparisonRequest' }
            example:
              tickers: [PETR4, VALE3]
              from: '2025-01-01'
              to: '2025-12-31'
              metrics: [return, volatility, drawdown]
              includeSeries: false
      responses:
        '200': { $ref: '#/components/responses/ComparisonSuccess' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '413':
          description: Corpo excede 64 KiB (request_too_large)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '415':
          description: Content-Type deve ser application/json (unsupported_media_type)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422': { $ref: '#/components/responses/InsufficientComparisonData' }
        '500': { $ref: '#/components/responses/InternalError' }
        '504': { $ref: '#/components/responses/RequestTimeout' }
  /quotes/{ticker}:
    get:
      operationId: getLatestQuote
      summary: Retorna a cotação mais recente publicada
      parameters:
        - $ref: '#/components/parameters/Ticker'
        - $ref: '#/components/parameters/MarketType'
      responses:
        '200':
          description: Cotação encontrada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QuoteEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /quotes/{ticker}/history:
    get:
      operationId: getQuoteHistory
      summary: Retorna cotações publicadas em um período
      description: Datas usam YYYY-MM-DD, com período máximo de cinco anos. O resultado contém apenas pregões e os preços são não ajustados.
      parameters:
        - $ref: '#/components/parameters/Ticker'
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
        - $ref: '#/components/parameters/MarketType'
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 1000, default: 100 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: order
          in: query
          schema: { type: string, enum: [asc, desc], default: asc }
      responses:
        '200':
          description: Página de cotações; data será vazio quando não houver pregões no período.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QuoteHistoryEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
components:
  parameters:
    Ticker:
      name: ticker
      in: path
      required: true
      schema: { type: string, pattern: '^[A-Za-z0-9]{4,12}$', example: PETR4 }
    From:
      name: from
      in: query
      required: true
      schema: { type: string, format: date, example: '2026-08-01' }
    To:
      name: to
      in: query
      required: true
      schema: { type: string, format: date, example: '2026-09-03' }
    MarketType:
      name: marketType
      in: query
      description: Código do tipo de mercado B3; 10 representa mercado à vista.
      schema: { type: integer, minimum: 1, default: 10 }
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [ok, unavailable] }
    OperationalMetrics:
      type: object
      required: [http, database, runtime]
      properties:
        http:
          type: array
          items:
            type: object
            required: [method, route, status, requests, total_duration_ms]
            properties:
              method: { type: string, enum: [GET, POST] }
              route: { type: string, example: '/assets/{ticker}/intelligence' }
              status: { type: integer, minimum: 100, maximum: 599 }
              requests: { type: integer, minimum: 0 }
              total_duration_ms: { type: number, minimum: 0 }
        database:
          type: array
          items:
            type: object
            required: [operation, calls, errors, rows, total_duration_ms]
            properties:
              operation: { type: string, enum: [readiness_ping, quote_latest, quote_history, comparison_history, intelligence_history] }
              calls: { type: integer, minimum: 0 }
              errors: { type: integer, minimum: 0 }
              rows: { type: integer, minimum: 0 }
              total_duration_ms: { type: number, minimum: 0 }
        runtime:
          type: object
          required: [goroutines, heap_alloc_bytes, heap_in_use_bytes, system_bytes, completed_gcs]
          properties:
            goroutines: { type: integer, minimum: 0 }
            heap_alloc_bytes: { type: integer, minimum: 0 }
            heap_in_use_bytes: { type: integer, minimum: 0 }
            system_bytes: { type: integer, minimum: 0 }
            completed_gcs: { type: integer, minimum: 0 }
    IntelligenceNumber:
      type: [number, 'null']
      description: Arredondado a duas casas; percentuais usam 20 para 20%. RSI usa escala 0 a 100.
    IntelligenceMoney:
      type: [string, 'null']
      pattern: '^-?[0-9]+\.[0-9]{2}$'
      description: Valor monetário em reais com duas casas decimais.
    IntelligenceSignal:
      type: object
      description: Avaliação técnica determinística. Não representa recomendação de investimento. Evidence é omitida quando status é unavailable.
      required: [id, status, severity]
      properties:
        id:
          type: string
          enum: [rsi_overbought, rsi_oversold, price_above_sma20, price_below_sma20, sma20_above_sma50, sma20_below_sma50]
        status: { type: string, enum: [triggered, not_triggered, unavailable] }
        severity: { type: string, enum: [info, warning] }
        evidence:
          type: object
          minProperties: 2
          properties:
            rsi14: { type: number, minimum: 0, maximum: 100, description: RSI efetivamente comparado pela regra. }
            threshold: { type: number, enum: [30, 70] }
            price: { type: number, description: Preço em reais efetivamente comparado pela regra. }
            sma20: { type: number, description: SMA20 em reais efetivamente comparada pela regra. }
            sma50: { type: number, description: SMA50 em reais efetivamente comparada pela regra. }
          example: { rsi14: 74.8, threshold: 70 }
    IntelligenceEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [ticker, price, returns, trend, momentum, risk, liquidity, signals]
          properties:
            ticker: { type: string, example: PETR4 }
            price:
              type: object
              required: [close, currency]
              properties:
                close: { $ref: '#/components/schemas/IntelligenceMoney' }
                currency: { type: string, const: BRL }
            returns:
              type: object
              required: [return_7d]
              properties:
                return_7d: { $ref: '#/components/schemas/IntelligenceNumber', description: Retorno de sete pregões; exige oito fechamentos. }
            trend:
              type: object
              required: [sma20, sma50, distance_sma20]
              properties:
                sma20: { $ref: '#/components/schemas/IntelligenceMoney', description: Média de vinte fechamentos; arredondada apenas na saída. }
                sma50: { $ref: '#/components/schemas/IntelligenceMoney', description: Média de cinquenta fechamentos; arredondada apenas na saída. }
                distance_sma20: { $ref: '#/components/schemas/IntelligenceNumber', description: Distância percentual da média não arredondada. }
            momentum:
              type: object
              required: [rsi14, rsi14_percentile]
              properties:
                rsi14: { $ref: '#/components/schemas/IntelligenceNumber', description: Wilder com semente nos primeiros quinze fechamentos publicados. }
                rsi14_percentile:
                  type: [object, 'null']
                  description: Percentil midrank do RSI atual contra valores anteriores à referência. Null quando houver menos de 252 observações válidas.
                  required: [value, window, observations, coverage]
                  properties:
                    value: { type: number, minimum: 0, maximum: 100, description: Arredondado a duas casas. }
                    window: { type: string, pattern: '^[1-9][0-9]*y$', example: 5y, description: Janela civil solicitada. }
                    observations: { type: integer, minimum: 252, example: 1034 }
                    coverage:
                      type: object
                      required: [from, to, complete]
                      properties:
                        from: { type: string, format: date, example: '2022-07-18' }
                        to: { type: string, format: date, example: '2026-09-08' }
                        complete: { type: boolean, example: false, description: Indica se havia histórico válido suficiente para alcançar o início civil da janela solicitada. }
            risk:
              type: object
              required: [volatility_30d, drawdown_current, maximum_drawdown_252d]
              properties:
                volatility_30d: { $ref: '#/components/schemas/IntelligenceNumber', description: Desvio amostral de trinta retornos simples anualizado por sqrt(252). }
                drawdown_current: { $ref: '#/components/schemas/IntelligenceNumber', description: Último fechamento contra o pico dos últimos 252 pregões. }
                maximum_drawdown_252d: { $ref: '#/components/schemas/IntelligenceNumber', description: Pior drawdown na janela de 252 pregões. }
            liquidity:
              type: object
              required: [average_daily_volume_20d]
              properties:
                average_daily_volume_20d: { $ref: '#/components/schemas/IntelligenceMoney', description: Volume financeiro médio de vinte pregões; fração de centavo truncada. }
            benchmark:
              type: object
              description: Presente somente quando benchmark é solicitado. Valores dependentes ficam null quando não há intervalo comparável.
              required: [ticker, returns, relative_strength]
              properties:
                ticker: { type: string, example: IBOV }
                returns:
                  type: object
                  required: [return_7d]
                  properties:
                    return_7d: { $ref: '#/components/schemas/IntelligenceNumber', description: Retorno do benchmark nas mesmas datas inicial e final usadas para a força relativa. }
                relative_strength:
                  type: object
                  required: [return_7d_pp]
                  properties:
                    return_7d_pp: { $ref: '#/components/schemas/IntelligenceNumber', description: Retorno percentual do ativo menos retorno percentual do benchmark, em pontos percentuais, sem arredondamento intermediário. }
            signals:
              type: array
              description: Todas as seis avaliações do ruleset 1.0, em ordem estável, incluindo estados not_triggered e unavailable.
              minItems: 6
              maxItems: 6
              items: { $ref: '#/components/schemas/IntelligenceSignal' }
        meta:
          type: object
          required: [status, market_type, as_of, source, price_adjustment, window_unit, percentage_unit, unavailable]
          properties:
            calendar:
              type: object
              description: Presente somente quando includeDetails=true.
              required: [source, version, policy, official_verified]
              properties:
                source: { type: string, example: cotahist_observed }
                version: { type: string, description: Hash da política, mercado, intervalo, cobertura e sessões. }
                policy: { type: string, example: observed_import_integrity_v1 }
                official_verified: { type: boolean, example: false }
                coverage:
                  type: array
                  items:
                    type: object
                    required: [year, observed_from, observed_to, import_integrity_validated, version]
                    properties:
                      year: { type: integer }
                      observed_from: { type: string, format: date }
                      observed_to: { type: string, format: date, description: Limite publicado do ano; pode ser posterior a asOf, mas sessões posteriores não são usadas. }
                      import_integrity_validated: { type: boolean }
                      version: { type: string }
            status: { type: string, enum: [complete, partial] }
            market_type: { type: integer, minimum: 1 }
            requested_as_of: { type: string, format: date, description: Presente somente quando asOf é informado. }
            as_of: { type: string, format: date }
            source: { type: string, const: B3 COTAHIST }
            price_adjustment: { type: string, const: unadjusted }
            window_unit: { type: string, const: trading_sessions }
            percentage_unit: { type: string, const: percent }
            calculation_version: { type: string, example: '1.2', description: Presente somente quando includeDetails=true. }
            ruleset_version: { type: string, const: '1.0', description: Presente somente quando includeDetails=true. Versão das regras determinísticas aplicadas aos indicadores. }
            data_version: { type: string, description: Presente somente quando includeDetails=true. Hash do histórico do ativo; quando há benchmark, combina as versões dos dois históricos. Não inclui calendário. }
            rsi_seed_from: { type: string, format: date, description: Presente somente quando includeDetails=true. }
            unavailable:
              type: array
              items:
                type: object
                required: [field, reason]
                properties:
                  field: { type: string, example: momentum.rsi14_percentile }
                  reason: { type: string, enum: [benchmark_not_found, insufficient_history, no_comparable_dates, calendar_unavailable, calendar_missing_year, calendar_invalid_coverage, calendar_outside_coverage, missing_sessions, invalid_data] }
    ComparisonMetrics:
      type: array
      description: Omissão ou lista JSON vazia aplica as métricas padrão; duplicidades são removidas.
      default: [return, volatility, drawdown, averageVolume, highLow, bestWorstDay]
      items:
        type: string
        enum: [return, volatility, drawdown, averageVolume, highLow, bestWorstDay, correlation]
    ComparisonRequest:
      type: object
      additionalProperties: false
      required: [tickers, from, to]
      properties:
        tickers:
          type: array
          minItems: 2
          maxItems: 10
          uniqueItems: true
          description: Entre 2 e 10 tickers únicos após normalização de espaços e letras maiúsculas.
          items: { type: string, example: PETR4 }
        from: { type: string, format: date }
        to: { type: string, format: date }
        marketType: { type: integer, minimum: 1, default: 10 }
        metrics: { $ref: '#/components/schemas/ComparisonMetrics' }
        benchmark:
          type: string
          description: Ticker consultado na mesma fonte. Não resolve identificadores de índices externos nem calcula excesso de retorno.
        includeSeries: { type: boolean, default: false }
    ComparisonEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [assets, ranking]
          properties:
            assets:
              type: array
              description: Preserva a ordem solicitada, incluindo ativos com dados insuficientes.
              items: { $ref: '#/components/schemas/AssetComparison' }
            ranking:
              type: object
              description: Ignora métricas indisponíveis; empates favorecem o primeiro ticker solicitado.
              properties:
                highestReturn: { type: string }
                lowestVolatility: { type: string }
                lowestDrawdown: { type: string, description: Ativo com drawdown mais próximo de zero. }
            correlations:
              type: array
              description: Pearson sem arredondamento adicional; pares indefinidos são omitidos.
              items:
                type: object
                required: [tickerA, tickerB, value]
                properties:
                  tickerA: { type: string }
                  tickerB: { type: string }
                  value: { type: number, minimum: -1, maximum: 1 }
            benchmark: { $ref: '#/components/schemas/AssetComparison' }
        meta:
          type: object
          required: [requestedTickers, from, to, marketType, metrics, priceAdjustment, currency, source, calculationVersion]
          properties:
            requestedTickers: { type: array, items: { type: string } }
            from: { type: string, format: date }
            to: { type: string, format: date }
            marketType: { type: integer }
            metrics: { $ref: '#/components/schemas/ComparisonMetrics' }
            priceAdjustment: { type: string, const: unadjusted }
            currency: { type: string, const: BRL }
            source: { type: string, const: B3 COTAHIST }
            calculationVersion: { type: string, const: '1.0' }
    AssetComparison:
      type: object
      required: [ticker, status, availableFrom, availableTo]
      description: Métricas não solicitadas ou indisponíveis são omitidas. Percentuais são arredondados a duas casas na apresentação.
      properties:
        ticker: { type: string }
        status: { type: string, enum: [available, insufficient_data] }
        availableFrom: { type: [string, 'null'], format: date }
        availableTo: { type: [string, 'null'], format: date }
        initialPrice: { $ref: '#/components/schemas/Money' }
        finalPrice: { $ref: '#/components/schemas/Money' }
        absoluteReturn: { $ref: '#/components/schemas/Money' }
        percentageReturn: { type: number }
        annualizedVolatility: { type: number, description: Desvio padrão amostral dos retornos multiplicado por sqrt(252) e 100. }
        maximumDrawdown: { type: number, maximum: 0 }
        averageDailyVolume: { $ref: '#/components/schemas/Money' }
        lowestPrice: { $ref: '#/components/schemas/Money' }
        highestPrice: { $ref: '#/components/schemas/Money' }
        bestDay: { $ref: '#/components/schemas/ComparisonDay' }
        worstDay: { $ref: '#/components/schemas/ComparisonDay' }
        message: { type: string }
        series:
          type: array
          description: Incluída apenas quando solicitada e o ativo possui dados suficientes.
          items:
            type: object
            required: [date, closePrice, normalizedPerformance]
            properties:
              date: { type: string, format: date }
              closePrice: { $ref: '#/components/schemas/Money' }
              normalizedPerformance: { type: number, description: Série com referência inicial igual a 100. }
    ComparisonDay:
      type: object
      required: [date, return]
      properties:
        date: { type: string, format: date }
        return: { type: number, description: Retorno percentual do pregão. }
    Quote:
      type: object
      required: [ticker, tradingDate, currency, marketType, openPrice, highPrice, lowPrice, closePrice]
      properties:
        ticker: { type: string, example: PETR4 }
        tradingDate: { type: string, format: date }
        bdiCode: { type: string }
        marketType: { type: integer }
        marketTypeDescription: { type: string, example: spot }
        shortName: { type: string }
        specification: { type: string, example: PN }
        term: { type: string }
        currency: { type: string, enum: [BRL] }
        openPrice: { $ref: '#/components/schemas/Money' }
        highPrice: { $ref: '#/components/schemas/Money' }
        lowPrice: { $ref: '#/components/schemas/Money' }
        averagePrice: { $ref: '#/components/schemas/Money' }
        closePrice: { $ref: '#/components/schemas/Money' }
        previousClose: { $ref: '#/components/schemas/Money' }
        absoluteChange: { $ref: '#/components/schemas/Money' }
        percentageChange: { type: [number, 'null'], format: double }
        intradayRange: { $ref: '#/components/schemas/Money' }
        bestBidPrice: { $ref: '#/components/schemas/Money' }
        bestAskPrice: { $ref: '#/components/schemas/Money' }
        bidAskSpread: { $ref: '#/components/schemas/Money' }
        tradeCount: { type: integer }
        tradedQuantity: { type: integer, format: int64 }
        tradedVolume: { $ref: '#/components/schemas/Money' }
        quoteFactor: { type: integer }
        isin: { type: string }
        distributionNumber: { type: integer }
    Money:
      type: string
      pattern: '^-?[0-9]+\.[0-9]{2}$'
      examples: ['48.20']
      description: Valor monetário decimal exato com duas casas, serializado como string.
    Metadata:
      type: object
      properties:
        ticker: { type: string }
        marketType: { type: integer }
        from: { type: string, format: date }
        to: { type: string, format: date }
        records: { type: integer }
        totalRecords: { type: integer, format: int64 }
        limit: { type: integer }
        offset: { type: integer }
        hasMore: { type: boolean }
        order: { type: string, enum: [asc, desc] }
        source: { type: string, const: B3 COTAHIST }
        sourceUrl: { type: string, format: uri }
        parserVersion: { type: string }
        layoutVersion: { type: string }
        publishedAt: { type: string, format: date-time }
        priceAdjustment: { type: string, const: unadjusted }
    QuoteEnvelope:
      type: object
      properties:
        data: { $ref: '#/components/schemas/Quote' }
        meta: { $ref: '#/components/schemas/Metadata' }
    QuoteHistoryEnvelope:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Quote' }
        meta: { $ref: '#/components/schemas/Metadata' }
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            field: { type: string }
            value: { type: string }
            expectedFormat: { type: string }
            retryable: { type: boolean }
  responses:
    ComparisonSuccess:
      description: Comparação calculada; pode conter ativos ou benchmark com dados insuficientes.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ComparisonEnvelope' }
    InsufficientComparisonData:
      description: Menos de dois ativos com dados suficientes (insufficient_comparison_data).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    InternalError:
      description: Erro interno sem exposição de detalhes do banco (internal_error).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RequestTimeout:
      description: Tempo limite da consulta excedido (request_timeout).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    BadRequest:
      description: Parâmetro inválido
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Cotação não encontrada
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }